// src/app/models/usuarioo.ts
export class Usuarioo {
    id_usuario: number = 0;
    nombre: string = "";
    apellido: string = "";
    fecha_creaci: string = "";
    correo: string = "";
    contrasena: string = "";
    status: number = 1;
    telefono: number = 0; 
    fecha_actual: string = "";
    genero : string = "h o m";
}